Fixed merge (duplicate files)
