+++ src/main/java/com/appdev/allin/playerData/PlayerDataRepo.java
@@ -1,11 +1,9 @@
 package com.appdev.allin.playerData;
 
-import java.time.LocalDate;
-import java.util.List;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
+import java.time.LocalDate;
+import java.util.List;
 import com.appdev.allin.player.Player;
 
 @Repository
